Rename old 1.1.1 Safe contract implementations (for Ethers.js + Web3.js)
